os-restart: don't exit prematurely if the kill fails
If you bounce a machine and then use the os-restart.sh script to bring
everything back up the initial `sudo pkill openshift` exits with
EXIT_FAILURE (because it's not actually running), and the -e on the
bash script will then exit immediately. This means that os-restart.sh
doesn't actually restart; fixed by moving the `set -error` to after
the `pkill openshift`.

Signed-off-by: Andrew McDermott <aim@frobware.com>